requirements.txt
boto3
bcrypt
python-dotenv